refactor(client): move BABYCARE enum patch behind the seam; hon_client is now _vendor-free
The HonParameterEnum BABYCARE monkeypatch (+ its lock/flag) moves out of
hon_client into client/pyhon_adapter.ensure_enum_patch (logic byte-identical;
only the relative import depth changes). hon_client.setup_sync calls it via the
adapter. Milestone: hon_client.py has NO _vendor imports anymore — the only
_vendor mentions left are comments + one log-message string. The single bridge
that imports _vendor is now client/pyhon_adapter.py.

test_enum_patch relocated to target the adapter (same stubbed namespace, same
idempotency/semantics/best-effort assertions, 6 passed). New ast guard
(test_session_adapter) blocks any future _vendor import from re-entering
hon_client. Refuter round: no findings. 119 passed, 1 skipped.
